backend/requirements.txt
fastapi
uvicorn[standard]
minio
python-dotenv
pydantic
pydantic-settings
python-multipart
sqlalchemy
psycopg2-binary
bcrypt==4.0.1
passlib[bcrypt]
python-jose[cryptography]